api/src/main/resources/requirements.txt
tornado-cors==0.6.0
requests[security]==2.9.1
python-swiftclient==3.0.0
python-keystoneclient==2.3.1
boto==2.40.0